Fix race condition in TrajectoryVisualization::incomingDisplayTrajectory
There is a data race on trajectory_message_to_display_ shared pointer.
In practice, this data race manifests itself as as TOCTOU bug.

One possible execution order that leads to a null pointer dereference:

- Thread A calls TrajectoryVisualization::update()
- Thread A locks update_trajectory_message_ mutex
- Thread A checks that trajectory_message_to_display_ is valid
- Thread B calls TrajectoryVisualization::incomingDisplayTrajectory
- Thread B invalidates the trajectory_message_to_display_ pointer by
  calling trajectory_message_to_display_.reset()
- Thread A tries to call trajectory_message_to_display_->getWayPointCount(),
  causing a null pointer dereference.

We fix the bug by protecting the call to
trajectory_message_to_display_.reset() with update_trajectory_message_
mutex, which was introduced in a related commit 149e5fb.